backend/requirements.txt
fastapi
uvicorn
google-api-python-client
openai
Pillow
pytesseract
aiohttp
python-dotenv
pydantic
regex
typing_extensions
python-multipart
